new data structure for closing

+++ src/hive_viriatum_commons/structures/linked_list.c
@@ -310,8 +310,9 @@ void popValueLinkedList(struct LinkedList_t *linkedList, void **valuePointer, ch
         *valuePointer = linkedListNode->value;
     }
 
-    /* in case the delete node flag is set */
-    if(deleteNode) {
+    /* in case the linked list node is valid
+    and the delete node flag is set */
+    if(linkedListNode != NULL && deleteNode) {
         /* deletes the linked list node */
         deleteLinkedListNode(linkedListNode);
     }
